item-service/requirements.txt
mysql-connector
grpcio
grpcio-tools